added logs and cleaned code
