+++ document_loader.py
@@ -4,8 +4,25 @@ from langchain_core.documents import Document
 import os
 from tqdm import tqdm
 import re
-import pathlib
-import fitz
+import logging
+import sys
+from datetime import datetime
+# Configure logging
+TODAY_DATE = datetime.now().strftime("%Y-%m-%d")
+LOG_DIR='Logs'
+os.makedirs(LOG_DIR, exist_ok=True)  # Ensure log directory exists
+LOG_FILE = os.path.join(LOG_DIR, f"logs_{TODAY_DATE}.log")
+logger = logging.getLogger(__name__)
+
+# Remove all existing handlers (if any)
+for handler in logger.handlers[:]:
+    logger.removeHandler(handler)
+    
+# File handler (writes logs to a file)
+file_handler = logging.FileHandler(LOG_FILE, encoding="utf-8")
+file_handler.setFormatter(logging.Formatter("%(asctime)s - %(levelname)s - %(filename)s - %(message)s"))
+logger.addHandler(file_handler)
+logger.propagate = False
 
 class DocLoader:
     """sumary_line
@@ -22,23 +39,25 @@ class DocLoader:
                 This can be particularly useful for metadata-driven pipelines or multi-format documents.
     """
     
-    def __init__(self,dirpath:str,filter: bool=False):
+    def __init__(self,dirpath:str,filter_text: bool=False):
         self.dirpath=dirpath
+        logger.info(f"Initializing DocLoader with directory: {dirpath}")
         self.file_path=self.load_pdf_files()
         self.__section_state = {"active_removal": False, "current_section": None}
-        self.filter = filter
+        self.filter_text = filter_text
 
     def load_pdf_files(self)-> list:
         """sumary_line
         Return: return list of path of pdf's in specific directory
         """
-        
+        logger.info("Loading PDF files from directory...")
         pdf_files = []
         for filename in tqdm(os.listdir(self.dirpath)):
             if filename.endswith(".pdf"):
                 filepath = os.path.join(self.dirpath, filename)
                 pdf_files.append(os.path.abspath(filepath))
         pdf_files = [p for p in pdf_files if os.path.isfile(p)]  # Validate paths
+        logger.info(f"Found {len(pdf_files)} PDF files.")
         return pdf_files
     
     def filter_sections(self, text: str) -> str:
@@ -51,6 +70,7 @@ class DocLoader:
         
         This function works line‐by‐line and uses persistent state across pages.
         """
+        logger.info("Filtering sections from extracted text.")
         # Define unwanted and preserved section names (all in lowercase)
         UNWANTED = [
             "references", "bibliography", "acknowledgments", "appendix",
@@ -80,6 +100,7 @@ class DocLoader:
             
             # Check if the candidate line starts with an unwanted header.
             if any(candidate.startswith(header) for header in UNWANTED):
+                logger.info(f"Ignoring section: {candidate}")
                 self.__section_state["active_removal"] = True
                 self.__section_state["current_section"] = candidate
                 # Do not include this header in output.
@@ -88,6 +109,7 @@ class DocLoader:
             # Check if the candidate line starts with a preserved header.
             if any(candidate.startswith(header) for header in PRESERVED):
                 # Stop removal mode.
+                logger.info(f"Preserving section: {candidate}")
                 self.__section_state["active_removal"] = False
                 self.__section_state["current_section"] = candidate
                 output_lines.append(line)  # Keep the header as-is.
@@ -102,131 +124,68 @@ class DocLoader:
         
         return "\n".join(output_lines).strip()
     
-    import re
-
-    def filter_sections2(self, text: str) -> str:
-        """
-        Removes unwanted sections from the text. Sections not in UNWANTED are preserved,
-        even if not explicitly in PRESERVED. Removal stops when any new section header (not in UNWANTED) is detected.
-        """
-        UNWANTED = [
-            "references", "bibliography", "acknowledgments", "appendix",
-            "conflicts of interest", "author contributions", "funding", 
-            "literature review", "introduction", "just accepted",
-            "supporting information", "references and notes","releated work","ethical consideration"
-        ]
-        PRESERVED = [
-            "abstract", "conclusion", "results and discussion", "results", "characterization",
-            "materials and methods", "methodology", "experiments", "conclusions", "keywords",
-            "experimental details", "experimental section","software discription","research contributions",
-            "proposed methodology","implementation","datasets","motivation and significance"
-        ]
-        SECTION_HEADER_PATTERN = re.compile(r'^\s*\d+[\.\-]?\s+.*')  # Matches "1. ", "2- ", etc.
-
-        if not hasattr(self, "__section_state"):
-            self.__section_state = {"active_removal": False}
-
-        output_lines = []
-        for line in text.splitlines():
-            stripped_line = line.strip()
-            # Detect section headers (numbered or from lists)
-            is_section_header = SECTION_HEADER_PATTERN.match(stripped_line) is not None
-
-            # Normalize candidate for comparison
-            candidate = stripped_line.lower()
-            candidate = re.sub(r"[:.,;]+$", "", candidate)
-            candidate = re.sub(r"^\d+\s*[\.\-]?\s*", "", candidate)
-
-            # Check if current line is an unwanted section
-            if any(candidate.startswith(header) for header in UNWANTED):
-                self.__section_state["active_removal"] = True
-                continue
-
-            # If it's a section header (numbered or preserved), stop removal
-            if is_section_header or any(candidate.startswith(header) for header in PRESERVED):
-                self.__section_state["active_removal"] = False
-
-            # Skip lines if in removal mode
-            if self.__section_state["active_removal"]:
-                continue
-
-            output_lines.append(line)
-
-        return "\n".join(output_lines).strip()
 
 
-
-
-
-
-
-
-
-
-
-    # def pypdf_loader(self):
-    #     """sumary_line
-    #     Simple and Fast document loader
-    #     """
-        
-    #     document=[]
-    #     ll =[self.file_path[0]]
-    #     for path in tqdm(ll):
-    #         pages =[]
-    #         loader = PyPDFLoader(file_path=path)
-    #         for page in loader.load():
-    #             # filtered_text = self.filter_sections(page.page_content)
-    #             # combined_content = "\n\n".join([page.page_content for page in pages])
-    #             pages.append(page)
-    #         document.append(pages)
-    #     return document
-    def pypdf_loader(self,filter=self.filter):
+    def pypdf_loader(self):
         """sumary_line
         Simple and Fast document loader
         """
-        
+        logger.info("Loading PDFs using PyPDFLoader...")
         document=[]
 
         for path in tqdm(self.file_path):
-            loader = PyPDFLoader(file_path=path)
-            combined_content =[]
-            metadata=None
-            for page in loader.load():
-                if metadata is None:
-                    metadata=page.metadata
-                    del metadata['page'],metadata['page_label']
-                    metadata.update({'source':os.path.basename(metadata['source'])})
-                combined_content.append(page.page_content)
+            try:
+                loader = PyPDFLoader(file_path=path)
+                combined_content =[]
+                metadata=None
+                for page in loader.load():
+                    if metadata is None:
+                        metadata=page.metadata
+                        del metadata['page'],metadata['page_label']
+                        metadata.update({'source':os.path.basename(metadata['source'])})
+                    combined_content.append(page.page_content)
 
-            combined_content = "\n\n".join(combined_content)
-            if filter:
-                document.append(Document(page_content=self.filter_sections(combined_content),metadata=metadata))
-            else:
-                document.append(Document(page_content=combined_content,metadata=metadata))
+                combined_content = "\n\n".join(combined_content)
+                if self.filter_text:
+                    document.append(Document(page_content=self.filter_sections(combined_content),metadata=metadata))
+                    logging.info(f"Processed file: {path}")
+                else:
+                    document.append(Document(page_content=combined_content,metadata=metadata))
+                    logging.info(f"Processed file: {path}")
+            except Exception as e:
+                logger.error(f"Error processing {path}: {e}")
+
+        logger.info(f"Total documents loaded: {len(document)}")      
         return document
 
-    def unstructured_loader(self,filter=self.filter):
+    def unstructured_loader(self):
         """_summary_
         better than simple pdf loader as it also includes images and other content which uses
         tesseract OCR so it takes more time than simple pdf loader
         """
+        logger.info("Loading PDFs using UnstructuredLoader...")
         document=[]
         for path in tqdm(self.file_path):
-            loader = UnstructuredLoader(file_path=path, strategy="hi_res")
-            fcombined_content =[]
-            metadata=None
-            for page in loader.load():
-                if metadata is None:
-                    metadata=page.metadata
-                    del metadata['page'],metadata['page_label']
-                    metadata.update({'source':os.path.basename(metadata['source'])})
-                combined_content.append(page.page_content)
+            try:
+                loader = UnstructuredLoader(file_path=path, strategy="hi_res")
+                fcombined_content =[]
+                metadata=None
+                for page in loader.load():
+                    if metadata is None:
+                        metadata=page.metadata
+                        del metadata['page'],metadata['page_label']
+                        metadata.update({'source':os.path.basename(metadata['source'])})
+                    combined_content.append(page.page_content)
 
-            combined_content = "\n\n".join(combined_content)
-            if filter:
-                document.append(Document(page_content=self.filter_sections(combined_content),metadata=metadata))
-            else:
-                document.append(Document(page_content=combined_content,metadata=metadata))
+                combined_content = "\n\n".join(combined_content)
+                if self.filter_text:
+                    document.append(Document(page_content=self.filter_sections(combined_content),metadata=metadata))
+                else:
+                    document.append(Document(page_content=combined_content,metadata=metadata))
+            except Exception as e:
+                logger.error(f"Error processing {path}: {e}")
+                
+        logger.info(f"Total documents loaded: {len(document)}")
         return document
 
     def docling_loader(self):
@@ -246,6 +205,7 @@ class DocLoader:
 
 
 if __name__=="__main__":
+    import pathlib
     obj = DocLoader("./PDF/")
     path=obj.load_pdf_files()
